requirements.txt
python-dotenv
discord-py-interactions
waybackpy